k1_extractor/requirements.txt
PyMuPDF>=1.23.0
pytesseract>=0.3.10
pdf2image>=1.16.3
openpyxl>=3.1.2
Pillow>=10.0.0
